indieauth/auth/requirements.txt
requests
pyjwt
mf2py